// apps/contact-api/src/common/enums/codes-error.enum.ts
export enum FormWidgetError {
    Project_No_Exists = "PROJECT_NO_EXISTS",
    Form_Not_Exists = "FORM_NOT_EXISTS",
    Error_Uuid_Not_Generated = "ERROR_UUID_NOT_GENERATED",
    Form_Exists = "FORM_EXISTS",
    Form_Not_Result_Search = "FORM_NOT_RESULT_SEARCH"
  };
  export enum GroupQuestionstError {
    Project_No_Exists = "PROJECT_NO_EXISTS",
    Group_Exists = "GROUP_EXISTS",
    Form_Not_Exists = "FORM_NOT_EXISTS",
    Group_Not_Result_Search = "GROUP_NOT_RESULT_SEARCH",
    Id_not_valid = "ID_NOT_VALID"
  };
  export enum ValidationFormError {
    Top_max_project = "MAX_FORM_PROJECT"
  };
  export enum SurveyErrorDto {
    Error_dto_create = "ERROR_DTO_CREATE",
    Error_not_result_search = "ERROR_NOT_RESULT_SEARCH"
  };