// apps/contact-api/src/common/enums/codes-success.enum.ts
export enum FormWidgetSuccess {
    Form_Create_Success = "FORM_CREATE_SUCCESS",
    Form_Delete_Success = "FORM_DELETE_SUCCESS",
    Form_Update_Success = "FORM_UPDATE_SUCCESS"
  };
  export enum GroupQuestionsSuccess {
    Group_Create_Success = "GROUP_CREATE_SUCCESS",
    Group_Update_Success = "GROUP_UPDATE_SUCCESS",
    Group_Delete_Success = "GROUP_DELETE_SUCCESS"
  };
  export enum SurveyGeneralSuccess {
    Survey_Create_Success = "SURVEY_CREATE_SUCCESS",
    Survey_Update_Success = "SURVEY_UPDATE_SUCCESS",
    Survey_Get_Success = "SURVEY_GET_SUCCESS"
  };